fix: use uri not path in explicit sample query set
Evaluations of the explicit sample query set currently return empty, because the `uri` of each target is in fact the path (e.g. `/foo/bar` instead of the Content ID (a UUID). Because the evaluations are empty, they fail to be uploaded to Prometheus, which is where we noticed the problem.