fix: refresh entire partition of evaluation inputs
The queries that append a new month's data to the tables in the
`automated_evaluation_input` dataset should wipe an existing partition
of the same date, rather than merge into it. The present `MERGE` can't
delete rows that should no longer be there, it can only appends rows
that aren't already there.

Fixes #43