fix: omit searches via query param includelist
SCH-1483

We need to evaluate the performance of a 'vanilla' search: the first
pages of results, in a default order, of an unfiltered, unredacted,
keyword-only search. We need the sample query sets of this evaluation to
omit searches that use any other filters, sort orders, or that request
results beyond the 10th page, or that have been redacted by substituting
any sensitive data with a label square brackets, such as `[email]`.

- Exclude searches whose query terms contain [ and ]. For example:
  [email], [postcode], or [postcode] council. These are not true
  queries, they’ve been redacted in our Google Analytics. So they’re
  meaningless in evaluations, because users are not searching for these
  literal queries.
- Exclude searches that use any parameters other than those allowed by
  the following rules.
  - Allow order=relevance, which is the default order
  - Allow parent, fbclid, (any others?)
  - Allow any parameter that doesn’t have a value
  - Allow page=n where n is less than or equal to 10

Implemented with the `LIKE` operator.

This is currently achieved by filtering out searches that involve
certain query parameters (an excludelist). The excludelist would need to
be updated whenever a new query parameter is defined.

The current excludelists don't include the query parameters `manual` or
`organisations`, which means that searches that use those query
parameters are incorrectly included in the evaluation sample query sets.

This commit replaces the excludelist with an includelist of query
parameters that can be used in a 'vanilla' search. The includelist will
only need to be updated when such parameters change, which will
probably be less often than when the parameters of other kinds of
searches change.

The other rules of inclusion and exclusion are specified alongside the
includelist.

A temporary function `UNPACK_PARAMS()` is defined to make the `WHERE`
clause more readable.