Make binary query sets 1:1 query:target
Fix #52

Context:

The purpose of the 'binary' evaluations is to have a score per target,
to test whether or not (hence, binary) the target appears in the search
results.

Currently, however, 'binary' evaluation scords are per query, not per
target. This is because each sampleQuery is constructed with one query
(the search string) and one or more targets. When this querySet is
evaluated, scores are calculated for the query, but not for each
individual target.

Change:

Construct each querySet with a single target. This will mean that the
evaluations, while still technically being evaluated per query, will
effectively be equivalent to being evaluated per target. If a given
target doesn't appear in the search results of a query, then the score
for that pair of query and target will be zero.