Match on attribution token for search events
Search events are only added to the search event tables
when there isn't already a matching row in the table. Whether
a row matches or not is determined by the "match on" criteria
in the SQL query. We already match on search term, timestamp,
user and results, but since the search term is a cleaned version
rather than the raw search term there are some edge cases where
two different searches would have different attribution tokens
but otherwise look the same in the search table. To handle these
edge cases, we will also match on the attribution token when
deciding whether a particular search already exists in the table
or not.